+++ app-sdk/utility/capture/include/frame_container.h
@@ -9,21 +9,22 @@
 #define FRAME_CONTAINER_H
 
 #include <macros.h>
+#include <stdint.h>
 
 class CAPTURE_EXPORT FrameContainer {
 public:
 	virtual ~FrameContainer() {
 	}
-	virtual size_t getEncodedLength() = 0;
+	virtual uint64_t getEncodedLength() = 0;
 	virtual int getWidth() = 0;
 	virtual int getHeight() = 0;
 	virtual int getChannels() = 0;
-	virtual size_t getFrameIndex() = 0;
-	virtual size_t getFrameCount() = 0;
+	virtual uint64_t getFrameIndex() = 0;
+	virtual uint64_t getFrameCount() = 0;
 	virtual bool isKey() = 0;
-	virtual bool decodeFrame(void*, size_t) = 0;
+	virtual bool decodeFrame(void*, uint64_t) = 0;
 	virtual void* data() = 0;
-	virtual size_t size() = 0;
+	virtual uint64_t size() = 0;
 };
 
 #endif /* VAFRAMECONTAINER_H_ */
